refact register_breakdown_handler
