+++ libs/enigo/Cargo.toml
@@ -37,8 +37,5 @@ core-graphics = "0.22"
 objc = "0.2"
 unicode-segmentation = "1.6"
 
-[target.'cfg(target_os = "linux")'.dependencies]
-libc = "0.2"
-
 [build-dependencies]
 pkg-config = "0.3"
