+++ src/core_main.rs
@@ -1,4 +1,4 @@
-use hbb_common::log;
+use hbb_common::{log, platform::register_breakdown_handler};
 
 /// shared by flutter and sciter main function
 ///
@@ -38,10 +38,10 @@ pub fn core_main() -> Option<Vec<String>> {
         }
         i += 1;
     }
+    register_breakdown_handler();
     #[cfg(target_os = "linux")]
     #[cfg(feature = "flutter")]
     {
-        crate::platform::linux::register_breakdown_handler();
         let (k, v) = ("LIBGL_ALWAYS_SOFTWARE", "true");
         if !hbb_common::config::Config::get_option("allow-always-software-render").is_empty() {
             std::env::set_var(k, v);
